+++ assets/src/components/flows/flow/FlowWorkbenchJobLauncher.tsx
@@ -1,6 +1,5 @@
 import { useClickOutside, useKeyDown } from '@react-hooks-library/core'
 import {
-  ArrowTopRightIcon,
   Button,
   Card,
   CloseIcon,
@@ -11,27 +10,20 @@ import {
   WorkbenchIcon,
 } from '@pluralsh/design-system'
 import { animated, useTransition } from '@react-spring/web'
-import chroma from 'chroma-js'
 import { GqlError } from 'components/utils/Alert'
 import { RectangleSkeleton } from 'components/utils/SkeletonLoaders'
-import { Body1P, Body2BoldP, Body2P } from 'components/utils/typography/Text'
+import { Body1P, Body2BoldP } from 'components/utils/typography/Text'
 import { WorkbenchJobCreateInput } from 'components/workbenches/workbench/WorkbenchJobCreateInput'
-import { POLL_INTERVAL } from 'components/cd/ContinuousDeployment'
 import {
   FlowBasicWithBindingsFragment,
-  useCancelWorkbenchJobMutation,
   useFlowWorkbenchesQuery,
-  useWorkbenchJobQuery,
   WorkbenchJobFragment,
   WorkbenchTinyFragment,
 } from 'generated/graphql'
-import { isJobRunning } from 'components/workbenches/workbench/job/WorkbenchJobActivity'
-import { Link } from 'react-router-dom'
-import { getWorkbenchJobAbsPath } from 'routes/workbenchesRoutesConsts'
-import { useCallback, useEffect, useMemo, useRef, useState } from 'react'
+import { useCallback, useMemo, useRef, useState } from 'react'
 import styled, { useTheme } from 'styled-components'
 import { isNonNullable } from 'utils/isNonNullable'
-import { RunStatusChip } from 'components/ai/infra-research/details/InfraResearch'
+import { WorkbenchStartedJobPanel } from 'components/workbenches/common/WorkbenchStartedJobPanel'
 
 export function FlowWorkbenchJobLauncher({
   flow,
@@ -101,15 +93,15 @@ function FlowWorkbenchJobPanel({
   const [createdJob, setCreatedJob] =
     useState<Nullable<WorkbenchJobFragment>>(null)
 
+  const selectedWorkbenchIdEffective =
+    selectedWorkbenchId || workbenches[0]?.id || ''
   const selectedWorkbench = useMemo(
-    () => workbenches.find((workbench) => workbench.id === selectedWorkbenchId),
-    [selectedWorkbenchId, workbenches]
+    () =>
+      workbenches.find(
+        (workbench) => workbench.id === selectedWorkbenchIdEffective
+      ),
+    [selectedWorkbenchIdEffective, workbenches]
   )
-
-  useEffect(() => {
-    if (selectedWorkbenchId && selectedWorkbench) return
-    setSelectedWorkbenchId(workbenches[0]?.id ?? '')
-  }, [selectedWorkbench, selectedWorkbenchId, workbenches])
 
   return (
     <Card
@@ -144,9 +136,10 @@ function FlowWorkbenchJobPanel({
       <PanelContentSC>
         {workbenchesError && <GqlError error={workbenchesError} />}
         {createdJob ? (
-          <CreatedWorkbenchJobContent
+          <WorkbenchStartedJobPanel
             initialJob={createdJob}
-            workbenchId={selectedWorkbenchId}
+            jobId={createdJob.id}
+            workbenchId={selectedWorkbenchIdEffective}
           />
         ) : (
           <>
@@ -156,7 +149,7 @@ function FlowWorkbenchJobPanel({
             >
               <Body2BoldP $color="text">Select workbench</Body2BoldP>
               <Select
-                selectedKey={selectedWorkbenchId}
+                selectedKey={selectedWorkbenchIdEffective}
                 isDisabled={!workbenchesLoading && !workbenches.length}
                 label={
                   workbenchesLoading ? (
@@ -186,9 +179,9 @@ function FlowWorkbenchJobPanel({
               <FlowContextSC />
             </Flex> */}
             <WorkbenchJobCreateInput
-              workbenchId={selectedWorkbenchId}
+              workbenchId={selectedWorkbenchIdEffective}
               workbenchLoading={workbenchesLoading}
-              disabled={!selectedWorkbenchId}
+              disabled={!selectedWorkbenchIdEffective}
               onCreated={setCreatedJob}
               placeholder="Start typing your question here..."
               wrapperStyles={{ minHeight: 140, maxWidth: '100%' }}
@@ -197,94 +190,6 @@ function FlowWorkbenchJobPanel({
         )}
       </PanelContentSC>
     </Card>
-  )
-}
-
-function CreatedWorkbenchJobContent({
-  initialJob,
-  workbenchId,
-}: {
-  initialJob: WorkbenchJobFragment
-  workbenchId: string
-}) {
-  const { data, startPolling, stopPolling } = useWorkbenchJobQuery({
-    variables: { id: initialJob.id },
-    fetchPolicy: 'cache-and-network',
-    pollInterval: POLL_INTERVAL,
-    skip: !initialJob.id,
-    notifyOnNetworkStatusChange: true,
-  })
-
-  const job = data?.workbenchJob ?? initialJob
-  const cancellable = isJobRunning(job.status)
-
-  useEffect(() => {
-    if (cancellable) startPolling(POLL_INTERVAL)
-    else stopPolling()
-    return () => stopPolling()
-  }, [cancellable, startPolling, stopPolling])
-
-  const [cancelWorkbenchJob, { loading, error }] =
-    useCancelWorkbenchJobMutation({
-      variables: { jobId: job.id },
-      awaitRefetchQueries: true,
-      refetchQueries: ['WorkbenchJob'],
-    })
-
-  return (
-    <Flex
-      direction="column"
-      gap="large"
-    >
-      {error && <GqlError error={error} />}
-      <StartedJobCardSC>
-        <IconFrame
-          type="secondary"
-          icon={<WorkbenchIcon color="icon-light" />}
-        />
-        <Flex
-          direction="column"
-          gap="xxsmall"
-          flex={1}
-          minWidth={0}
-        >
-          <Body2BoldP $color="text">Started workbench job</Body2BoldP>
-          <Body2P
-            $color="text-light"
-            css={{
-              display: '-webkit-box',
-              WebkitLineClamp: 3,
-              WebkitBoxOrient: 'vertical',
-              overflow: 'hidden',
-            }}
-          >
-            {job.prompt}
-          </Body2P>
-        </Flex>
-        <RunStatusChip
-          status={job.status}
-          showSpinner
-        />
-      </StartedJobCardSC>
-      <Flex justify={cancellable ? 'space-between' : 'flex-end'}>
-        {cancellable && (
-          <Button
-            destructive
-            loading={loading}
-            onClick={() => cancelWorkbenchJob()}
-          >
-            Cancel workbench run
-          </Button>
-        )}
-        <Button
-          as={Link}
-          to={getWorkbenchJobAbsPath({ workbenchId, jobId: job.id })}
-          endIcon={<ArrowTopRightIcon />}
-        >
-          See progress
-        </Button>
-      </Flex>
-    </Flex>
   )
 }
 
@@ -305,30 +210,6 @@ const PanelContentSC = styled(Flex)(({ theme }) => ({
   marginTop: theme.spacing.medium,
   minHeight: 0,
   overflow: 'visible',
-}))
-
-const StartedJobCardSC = styled.div(({ theme }) => ({
-  alignItems: 'center',
-  background: theme.colors['fill-two'],
-  borderRadius: theme.borderRadiuses.medium,
-  display: 'flex',
-  gap: theme.spacing.medium,
-  padding: theme.spacing.medium,
-  position: 'relative',
-
-  '&::before': {
-    content: '""',
-    position: 'absolute',
-    inset: 0,
-    borderRadius: 'inherit',
-    padding: 1,
-    background: `linear-gradient(315deg, ${chroma('#5C77FF').alpha(0).hex()} 0%, #494FF2 46%, ${chroma('#8FD6FF').alpha(0.6).hex()} 79%, ${chroma('#52F4D9').alpha(0.5).hex()} 100%)`,
-    pointerEvents: 'none',
-    WebkitMask:
-      'linear-gradient(#fff 0 0) content-box, linear-gradient(#fff 0 0)',
-    WebkitMaskComposite: 'xor',
-    maskComposite: 'exclude',
-  },
 }))
 
 // const FlowContextSC = styled.div(({ theme }) => ({
