+++ assets/src/components/workbenches/common/WorkbenchEvalSkillButton.tsx
@@ -1,6 +1,5 @@
 import { useClickOutside, useKeyDown } from '@react-hooks-library/core'
 import {
-  ArrowTopRightIcon,
   Button,
   Card,
   CloseIcon,
@@ -9,24 +8,13 @@ import {
   WorkbenchIcon,
 } from '@pluralsh/design-system'
 import { animated, useTransition } from '@react-spring/web'
-import chroma from 'chroma-js'
 import { ChatInputSimple } from 'components/ai/chatbot/input/ChatInput'
 import { GqlError } from 'components/utils/Alert'
-import { RectangleSkeleton } from 'components/utils/SkeletonLoaders'
-import { useSimpleToast } from 'components/utils/SimpleToastContext'
-import { Body1P, Body2BoldP, Body2P } from 'components/utils/typography/Text'
-import { POLL_INTERVAL } from 'components/cd/ContinuousDeployment'
-import {
-  useCancelWorkbenchJobMutation,
-  useWorkbenchEvalSkillMutation,
-  useWorkbenchJobQuery,
-} from 'generated/graphql'
-import { isJobRunning } from 'components/workbenches/workbench/job/WorkbenchJobActivity'
-import { RunStatusChip } from 'components/ai/infra-research/details/InfraResearch'
-import { ComponentProps, useCallback, useEffect, useRef, useState } from 'react'
-import { Link } from 'react-router-dom'
-import { getWorkbenchJobAbsPath } from 'routes/workbenchesRoutesConsts'
+import { Body1P } from 'components/utils/typography/Text'
+import { useWorkbenchEvalSkillMutation } from 'generated/graphql'
+import { ComponentProps, useCallback, useRef, useState } from 'react'
 import styled, { useTheme } from 'styled-components'
+import { WorkbenchStartedJobPanel } from 'components/workbenches/common/WorkbenchStartedJobPanel'
 
 type WorkbenchEvalSkillButtonProps = Omit<
   ComponentProps<typeof Button>,
@@ -45,21 +33,25 @@ export function WorkbenchEvalSkillButton({
   const theme = useTheme()
   const ref = useRef<HTMLDivElement>(null)
   const lastEvalResultIdRef = useRef<Nullable<string>>(evalResultId ?? null)
-  const { popToast } = useSimpleToast()
   const [open, setOpen] = useState(false)
   const [additionalInformation, setAdditionalInformation] = useState('')
   const [createdJobId, setCreatedJobId] = useState<Nullable<string>>(null)
+  const [mutationError, setMutationError] = useState<Nullable<Error>>(null)
 
   const [workbenchEvalSkill, { loading }] = useWorkbenchEvalSkillMutation({
     onCompleted: ({ workbenchEvalSkill }) => {
-      popToast({
-        content: 'Skills updated successfully',
-        severity: 'success',
-      })
+      setMutationError(null)
       setCreatedJobId(workbenchEvalSkill?.id ?? null)
     },
-    onError: (e) => popToast({ content: e.message, severity: 'danger' }),
+    onError: (e) => setMutationError(e),
   })
+
+  const resetState = useCallback(() => {
+    setOpen(false)
+    setCreatedJobId(null)
+    setMutationError(null)
+    setAdditionalInformation('')
+  }, [])
 
   const toggle = useCallback(() => {
     if (!open && lastEvalResultIdRef.current !== (evalResultId ?? null)) {
@@ -79,7 +71,7 @@ export function WorkbenchEvalSkillButton({
   })
 
   useKeyDown(['Escape'], () => setOpen(false))
-  useClickOutside(ref, () => setOpen(false))
+  useClickOutside(ref, resetState)
 
   return (
     <div
@@ -98,9 +90,10 @@ export function WorkbenchEvalSkillButton({
           <WorkbenchEvalSkillPanel
             additionalInformation={additionalInformation}
             createdJobId={createdJobId}
+            mutationError={mutationError}
             loading={loading}
             onAdditionalInformationChange={setAdditionalInformation}
-            onClose={() => setOpen(false)}
+            onClose={resetState}
             onSubmit={() => {
               if (!evalResultId || loading) return
 
@@ -122,6 +115,7 @@ export function WorkbenchEvalSkillButton({
 function WorkbenchEvalSkillPanel({
   additionalInformation,
   createdJobId,
+  mutationError,
   loading,
   onAdditionalInformationChange,
   onClose,
@@ -130,6 +124,7 @@ function WorkbenchEvalSkillPanel({
 }: {
   additionalInformation: string
   createdJobId: Nullable<string>
+  mutationError: Nullable<Error>
   loading: boolean
   onAdditionalInformationChange: (value: string) => void
   onClose: () => void
@@ -169,8 +164,9 @@ function WorkbenchEvalSkillPanel({
         />
       </Flex>
       <PanelContentSC>
+        {mutationError && <GqlError error={mutationError} />}
         {createdJobId ? (
-          <CreatedWorkbenchJobContent
+          <WorkbenchStartedJobPanel
             jobId={createdJobId}
             workbenchId={workbenchId}
           />
@@ -195,108 +191,6 @@ function WorkbenchEvalSkillPanel({
   )
 }
 
-function CreatedWorkbenchJobContent({
-  jobId,
-  workbenchId,
-}: {
-  jobId: string
-  workbenchId: string
-}) {
-  const {
-    data,
-    error: queryError,
-    startPolling,
-    stopPolling,
-  } = useWorkbenchJobQuery({
-    variables: { id: jobId },
-    fetchPolicy: 'cache-and-network',
-    pollInterval: POLL_INTERVAL,
-    skip: !jobId,
-    notifyOnNetworkStatusChange: true,
-  })
-
-  const job = data?.workbenchJob
-  const cancellable = job ? isJobRunning(job.status) : false
-
-  useEffect(() => {
-    if (cancellable) startPolling(POLL_INTERVAL)
-    else stopPolling()
-
-    return () => stopPolling()
-  }, [cancellable, startPolling, stopPolling])
-
-  const [cancelWorkbenchJob, { loading, error: cancelError }] =
-    useCancelWorkbenchJobMutation({
-      variables: { jobId },
-      awaitRefetchQueries: true,
-      refetchQueries: ['WorkbenchJob'],
-    })
-
-  return (
-    <Flex
-      direction="column"
-      gap="large"
-    >
-      {queryError && <GqlError error={queryError} />}
-      {cancelError && <GqlError error={cancelError} />}
-      {job ? (
-        <StartedJobCardSC>
-          <IconFrame
-            type="secondary"
-            icon={<WorkbenchIcon color="icon-light" />}
-          />
-          <Flex
-            direction="column"
-            gap="xxsmall"
-            flex={1}
-            minWidth={0}
-          >
-            <Body2BoldP $color="text">Started workbench job</Body2BoldP>
-            <Body2P
-              $color="text-light"
-              css={{
-                display: '-webkit-box',
-                WebkitLineClamp: 3,
-                WebkitBoxOrient: 'vertical',
-                overflow: 'hidden',
-              }}
-            >
-              {job.prompt}
-            </Body2P>
-          </Flex>
-          <RunStatusChip
-            status={job.status}
-            showSpinner
-          />
-        </StartedJobCardSC>
-      ) : (
-        <RectangleSkeleton
-          $height={112}
-          $width="100%"
-        />
-      )}
-      <Flex justify={cancellable ? 'space-between' : 'flex-end'}>
-        {cancellable && (
-          <Button
-            destructive
-            loading={loading}
-            onClick={() => cancelWorkbenchJob()}
-          >
-            Cancel workbench run
-          </Button>
-        )}
-        <Button
-          as={Link}
-          to={getWorkbenchJobAbsPath({ workbenchId, jobId })}
-          endIcon={<ArrowTopRightIcon />}
-        >
-          See progress
-        </Button>
-      </Flex>
-    </Flex>
-  )
-}
-
 const AnimatedWrapperSC = styled(animated.div)(({ theme }) => ({
   position: 'absolute',
   right: 0,
@@ -314,28 +208,4 @@ const PanelContentSC = styled(Flex)(({ theme }) => ({
   marginTop: theme.spacing.medium,
   minHeight: 0,
   overflow: 'visible',
-}))
-
-const StartedJobCardSC = styled.div(({ theme }) => ({
-  alignItems: 'center',
-  background: theme.colors['fill-two'],
-  borderRadius: theme.borderRadiuses.medium,
-  display: 'flex',
-  gap: theme.spacing.medium,
-  padding: theme.spacing.medium,
-  position: 'relative',
-
-  '&::before': {
-    content: '""',
-    position: 'absolute',
-    inset: 0,
-    borderRadius: 'inherit',
-    padding: 1,
-    background: `linear-gradient(315deg, ${chroma('#5C77FF').alpha(0).hex()} 0%, #494FF2 46%, ${chroma('#8FD6FF').alpha(0.6).hex()} 79%, ${chroma('#52F4D9').alpha(0.5).hex()} 100%)`,
-    pointerEvents: 'none',
-    WebkitMask:
-      'linear-gradient(#fff 0 0) content-box, linear-gradient(#fff 0 0)',
-    WebkitMaskComposite: 'xor',
-    maskComposite: 'exclude',
-  },
 }))
